More keep cleanup, got things to compile
Next: style changes
